fix: change hscode to ncm

+++ Sinerlog.Lambda.Pdf.Label/Application/LabelGenerator.cs
@@ -57,7 +57,7 @@ namespace Sinerlog.Lambda.Pdf.Label.Application
             texto = texto.Replace("{receiverCity}", label.ReceiverCity);
             texto = texto.Replace("{receiverState}", label.ReceiverState);
             texto = texto.Replace("{serviceType}", label.ServiceType);
-            texto = texto.Replace("{weight}", label.Weight.ToString());
+            texto = texto.Replace("{weight}", ((int)label.Weight).ToString());
 
             return texto;
         }
